Use a source_set rather than a static_library
Some toolchains (e.g., the host toolchain on Mac OS X) cannot build this
code as a static_library because it does not contain any complication
units (and hence the .a would be empty). Instead, we can model this
library as a source_set, which both works and is more accurate.

Change-Id: Ib11e866b74297e1a7c8c5a51312d3945bf154788